Add "featured trips" to home page
Featured trips have been added to the home page. The featured trips are three randomly selected trips that are also able to be added to the cart. LocalStorage has been used to check the "cart-state" to ensure that cruises added from the home screen, end up in the cart too.